requirements.txt
fastapi
uvicorn
transformers
datasets
torchaudio
pydub
soundfile
ffmpeg-python
python-multipart
coqui-tts
pytest
httpx
